Update gulp to have custom beeping errors
This appears to work in cmd but not git shell or babun!

gulp-plumber: this attaches an error handler to a stream, so that you
don't have to do .on("error") on every step in the stream
gulp-beepbeep: plays a beep sound in the console

When a SASS or JS error is encountered, the error is logged in colors
and a beep is triggered.