Fix: polygon light logic update
The calculation for polygon light position was borked by the caching
light update. In the process, I decided to slightly optimize. Instead of
each polygon's canvas being equal to the game window size, I shrunk it
down to be just the size needed to fit any rotation of said polygon